+++ lib/theme.js
@@ -1,0 +1,19 @@
+// @flow
+const colors = {
+  dusty: 'rgba(217, 229, 214, 1);',
+  blue: 'rgba(0, 167, 225, 1);',
+  warm: 'rgba(237, 222, 164, 1);',
+  peach: 'rgba(247, 160, 114, 1);',
+  orange: 'rgba(255, 155, 66, 1);',
+};
+
+const sizes = {
+  lineWidth: '3px',
+};
+
+const theme: Brew$ThemeType = {
+  colors,
+  sizes,
+};
+
+export default theme;
