+++ pages/index.js
@@ -12,18 +12,7 @@ import StepWrapper from '../components/StepWrapper';
 
 import BrewStep from '../components/BrewStep';
 import { TimerContextProvider } from '../components/Timer/Timer';
-
-const colors = {
-  dusty: 'rgba(217, 229, 214, 1);',
-  blue: 'rgba(0, 167, 225, 1);',
-  warm: 'rgba(237, 222, 164, 1);',
-  peach: 'rgba(247, 160, 114, 1);',
-  orange: 'rgba(255, 155, 66, 1);',
-};
-
-const sizes = {
-  lineWidth: '3px',
-};
+import theme from '../lib/theme';
 
 type State = {
   activeStep: 'weight' | 'profile' | 'reset' | 'brew',
@@ -87,7 +76,7 @@ class Index extends PureComponent<*, State> {
       taste,
     } = this.state;
     return (
-      <ThemeProvider theme={{ colors, sizes }}>
+      <ThemeProvider theme={theme}>
         <Page>
           <Content>
             <StepWrapper forwardRef={this.stepRefs.amountStep} isActive>
